+++ tests/test_pylast_library.py
@@ -1,0 +1,30 @@
+#!/usr/bin/env python
+"""
+Integration (not unit) tests for pylast.py
+"""
+import unittest
+
+import pylast
+
+from .test_pylast import PyLastTestCase
+
+
+class TestPyLastLibrary(PyLastTestCase):
+
+    def test_library_is_hashable(self):
+        # Arrange
+        library = pylast.Library(user=self.username, network=self.network)
+
+        # Act/Assert
+        self.helper_is_thing_hashable(library)
+
+    def test_cacheable_library(self):
+        # Arrange
+        library = pylast.Library(self.username, self.network)
+
+        # Act/Assert
+        self.helper_validate_cacheable(library, "get_artists")
+
+
+if __name__ == '__main__':
+    unittest.main(failfast=True)
